Options class added
This class is basically used to load the plugin options from a YAML
file.
It has one function, load(String option, Class optionClass) that tries
to load the YAML file into a HashMap,  and checks if it cointains the
entry <option> of the class <optionClass> and returns false if not.